Module nrf52::clock

The clock peripheral of the nRF51 series (chapter 13 of the nRF51 reference manual v3.0), which manages the low frequency and high frequency clocks. The low frequency clock drives the real time clock (RTC), while the high frequency clocks drive the timer system.

Author: Philip Levis Date: August 18, 2016

Reexports

use core::cell::Cell;

use core::mem;

use kernel::common::VolatileCell;

Structs

Clock

Registers

Enums

ClockRunning

ClockTaskTriggered

HighClockSource

InterruptField

LowClockSource

XtalFreq

Constants

CLOCK_BASE

Statics

CLOCK

Traits

ClockClient

Functions

CLK