Module nrf52::gpio

FIXME: CHECK FOR nRF52-DK

GPIO and GPIOTE (task and events) for the nRF51 series.

Author: Philip Levis pal@cs.stanford.edu Date: August 18, 2016

Reexports

use chip;

use core::cell::Cell;

use core::mem;

use core::ops::Index;

use core::ops::IndexMut;

use kernel::common::VolatileCell;

use kernel::hil;

use nvic;

use peripheral_interrupts::NvicIdx;

use peripheral_registers::GPIO_BASE;

use peripheral_registers::GPIO;

Structs

GPIOPin

GpioteRegisters
The nRF51822 doesn't automatically provide GPIO interrupts. Instead, to receive interrupts from a GPIO line, you must allocate a GPIOTE (GPIO Task and Event) channel, and bind the channel to the desired pin. There are 4 channels. This means that requesting an interrupt can fail, if there are already 4 allocated.

Port

Constants

GPIOTE_BASE

Statics

PORT

Functions

GPIO

GPIOTE

GPIOTE_Handler⚠

allocate_channel
Allocate a GPIOTE channel

find_channel
Return which channel is allocated to a pin, or -1 if none.