Struct nrf52::gpio::GpioteRegisters

#[repr(C, packed)]

struct GpioteRegisters {
    pub out0: VolatileCell<u32>,
    pub out1: VolatileCell<u32>,
    pub out2: VolatileCell<u32>,
    pub out3: VolatileCell<u32>,
    _reserved0: [VolatileCell<u8>; 240],
    pub in0: VolatileCell<u32>,
    pub in1: VolatileCell<u32>,
    pub in2: VolatileCell<u32>,
    pub in3: VolatileCell<u32>,
    _reserved1: [VolatileCell<u8>; 108],
    pub port: VolatileCell<u32>,
    _reserved2: [VolatileCell<u8>; 384],
    pub inten: VolatileCell<u32>,
    pub intenset: VolatileCell<u32>,
    pub intenclr: VolatileCell<u32>,
    _reserved3: [VolatileCell<u8>; 516],
    pub config0: VolatileCell<u32>,
    pub config1: VolatileCell<u32>,
    pub config2: VolatileCell<u32>,
    pub config3: VolatileCell<u32>,
}

The nRF51822 doesn't automatically provide GPIO interrupts. Instead, to receive interrupts from a GPIO line, you must allocate a GPIOTE (GPIO Task and Event) channel, and bind the channel to the desired pin. There are 4 channels. This means that requesting an interrupt can fail, if there are already 4 allocated.

Fields
out0: VolatileCell<u32> out1: VolatileCell<u32> out2: VolatileCell<u32> out3: VolatileCell<u32> _reserved0: [VolatileCell<u8>; 240] in0: VolatileCell<u32> in1: VolatileCell<u32> in2: VolatileCell<u32> in3: VolatileCell<u32> _reserved1: [VolatileCell<u8>; 108] port: VolatileCell<u32> _reserved2: [VolatileCell<u8>; 384] inten: VolatileCell<u32> intenset: VolatileCell<u32> intenclr: VolatileCell<u32> _reserved3: [VolatileCell<u8>; 516] config0: VolatileCell<u32> config1: VolatileCell<u32> config2: VolatileCell<u32> config3: VolatileCell<u32>