Struct nrf52::timer::Timer

pub struct Timer {
    which: Location,
    nvic: NvicIdx,
    client: Cell<Option<&'static CompareClient>>,
}

Fields
which: Location nvic: NvicIdx client: Cell<Option<&'static CompareClient>>
Methods

impl Timer

fn timer(&self) -> &'static TIMER

const fn new(location: Location, nvic: NvicIdx) -> Timer

fn set_client(&self, client: &'static CompareClient)

fn start(&self)

fn stop(&self)

fn shutdown(&self)

fn clear(&self)

fn capture(&self, which: u8) -> u32

Capture the current timer value into the CC register specified by which, and return the value.

fn capture_to(&self, which: u8)

Capture the current value to the CC register specified by which and do not return the value.

fn get_shortcuts(&self) -> u32

Shortcuts can automatically stop or clear the timer on a particular compare event; refer to section 18.3 of the nRF reference manual for details. Implementation currently provides shortcuts as the raw bitmask.

fn set_shortcuts(&self, shortcut: u32)

fn get_cc0(&self) -> u32

fn set_cc0(&self, val: u32)

fn get_cc1(&self) -> u32

fn set_cc1(&self, val: u32)

fn get_cc2(&self) -> u32

fn set_cc2(&self, val: u32)

fn get_cc3(&self) -> u32

fn set_cc3(&self, val: u32)

fn enable_interrupts(&self, interrupts: u32)

fn disable_interrupts(&self, interrupts: u32)

fn enable_nvic(&self)

fn disable_nvic(&self)

fn set_prescaler(&self, val: u8)

fn get_prescaler(&self) -> u8

fn handle_interrupt(&self)

When an interrupt occurs, check if any of the 4 compares have created an event, and if so, add it to the bitmask of triggered events that is passed to the client.